Module aws_sdk_connectparticipant::error

Modules

access_denied_exception

See AccessDeniedException

conflict_exception

See ConflictException

internal_server_exception

See InternalServerException

service_quota_exceeded_exception

See ServiceQuotaExceededException

throttling_exception

See ThrottlingException

validation_exception

See ValidationException

Structs

AccessDeniedException

You do not have sufficient access to perform this action.

CompleteAttachmentUploadError

ConflictException

An attachment with that identifier is already being uploaded.

CreateParticipantConnectionError

DisconnectParticipantError

GetAttachmentError

GetTranscriptError

InternalServerException

This exception occurs when there is an internal failure in the Amazon Connect service.

SendEventError

SendMessageError

ServiceQuotaExceededException

The number of attachments per contact exceeds the quota.

StartAttachmentUploadError

ThrottlingException

The request was denied due to request throttling.

ValidationException

The input fails to satisfy the constraints specified by Amazon Connect.

Enums

CompleteAttachmentUploadErrorKind

CreateParticipantConnectionErrorKind

DisconnectParticipantErrorKind

GetAttachmentErrorKind

GetTranscriptErrorKind

SendEventErrorKind

SendMessageErrorKind

StartAttachmentUploadErrorKind