Module aws_sdk_connectparticipant::operation

Structs

CompleteAttachmentUpload

Allows you to confirm that the attachment has been uploaded using the pre-signed URL

CreateParticipantConnection

Creates the participant's connection. Note that ParticipantToken is used for invoking

DisconnectParticipant

Disconnects a participant. Note that ConnectionToken is used for invoking this API

GetAttachment

Provides a pre-signed URL for download of a completed attachment. This is an

GetTranscript

Retrieves a transcript of the session, including details about any attachments. Note

SendEvent

Sends an event. Note that ConnectionToken is used for invoking this API instead of

SendMessage

Sends a message. Note that ConnectionToken is used for invoking this API instead of

StartAttachmentUpload

Provides a pre-signed Amazon S3 URL in response for uploading the file directly to